worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
feedparser==6.0.11
pydantic==2.9.2
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
httpx==0.27.2
python-dotenv==1.0.0
requests==2.31.0
json5==0.9.25
openai==1.40.0
langchain_community==0.2.17
langchain==0.2.17
